Listing all urls

+++ Sources/CloudDocs/CloudDocs.swift
@@ -16,6 +16,7 @@ public struct CloudDocs {
         case fileAllreadyExists
         case fileNotFound
         case cloudDocumentFolderNotFound
+        case urlsNotFound
     }
 }
 
@@ -34,25 +35,32 @@ public extension CloudDocs {
     func removeFile(fileName: String,
                     fileExtension: String? = nil,
                     completion: @escaping (_ success: Bool, _ error: Error?) -> Void) {
-        DispatchQueue.global().async { [self] in
-            guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
-                completion(false, CloudDocsError.cloudDocumentFolderNotFound)
+        guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
+            completion(false, CloudDocsError.cloudDocumentFolderNotFound)
+            return
+        }
+        _listAllFileURLs(from: cloudDocumentContainerUrl) { (urls: [URL]?, error: Error?) in
+            if let error = error {
+                completion(false, error)
                 return
             }
-            do {
-                let urls = try fileManager.contentsOfDirectory(at: cloudDocumentContainerUrl,
-                                                               includingPropertiesForKeys: nil,
-                                                               options: [])
-                let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
-                if urls.contains(fileURL) {
+            guard let urls = urls else {
+                completion(false, CloudDocsError.urlsNotFound)
+                return
+            }
+            let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
+            if urls.contains(fileURL) {
+                do {
                     try fileManager.removeItem(at: fileURL)
                     completion(true, nil)
-                } else {
-                    completion(false, CloudDocsError.fileNotFound)
+                    return
+                } catch {
+                    completion(false, error)
                     return
                 }
-            } catch {
-                completion(false, error)
+            } else {
+                completion(false, CloudDocsError.fileNotFound)
+                return
             }
         }
     }
@@ -60,27 +68,32 @@ public extension CloudDocs {
     func readFile<File: Decodable>(fileName: String,
                                    fileExtension: String? = nil,
                                    completion: @escaping (_ file: File?, _ error: Error?) -> Void) {
-        DispatchQueue.global().async { [self] in
-            guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
-                completion(nil, CloudDocsError.cloudDocumentFolderNotFound)
+        guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
+            completion(nil, CloudDocsError.cloudDocumentFolderNotFound)
+            return
+        }
+        _listAllFileURLs(from: cloudDocumentContainerUrl) { (urls: [URL]?, error: Error?) in
+            if let error = error {
+                completion(nil, error)
                 return
             }
-            do {
-                let urls = try fileManager.contentsOfDirectory(at: cloudDocumentContainerUrl,
-                                                               includingPropertiesForKeys: nil,
-                                                               options: [])
-                let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
-                if urls.contains(fileURL) {
+            guard let urls = urls else {
+                completion(nil, CloudDocsError.urlsNotFound)
+                return
+            }
+            let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
+            if urls.contains(fileURL) {
+                do {
                     let data = try Data(contentsOf: URL(fileURLWithPath: fileURL.path), options: .mappedIfSafe)
                     let jsonResult = try JSONDecoder().decode(File.self, from: data)
                     completion(jsonResult, nil)
                     return
-                } else {
-                    completion(nil, CloudDocsError.fileNotFound)
+                } catch {
+                    completion(nil, error)
                     return
                 }
-            } catch {
-                completion(nil, error)
+            } else {
+                completion(nil, CloudDocsError.fileNotFound)
                 return
             }
         }
@@ -91,36 +104,46 @@ public extension CloudDocs {
                                         content: Content,
                                         force: Bool = false,
                                         completion: @escaping (_ success: Bool, _ error: Error?) -> Void) {
-        DispatchQueue.global().async { [self] in
-            guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
-                completion(false, CloudDocsError.cloudDocumentFolderNotFound)
+        guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
+            completion(false, CloudDocsError.cloudDocumentFolderNotFound)
+            return
+        }
+        _listAllFileURLs(from: cloudDocumentContainerUrl) { (urls: [URL]?, error: Error?) in
+            if let error = error {
+                completion(false, error)
                 return
             }
-            createFolderIfNotExists(from: cloudDocumentContainerUrl) { (error: Error?) in
-                if let error = error {
+            guard let urls = urls else {
+                completion(false, CloudDocsError.urlsNotFound)
+                return
+            }
+            let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
+            if !urls.contains(fileURL) || force {
+                let encoder = JSONEncoder()
+                encoder.outputFormatting = .prettyPrinted
+                do {
+                    let encodedContent = try encoder.encode(content)
+                    let created = fileManager.createFile(atPath: fileURL.path,
+                                                         contents: encodedContent, attributes: nil)
+                    completion(created, nil)
+                    return
+                } catch {
                     completion(false, error)
                     return
                 }
-                do {
-                    let urls = try fileManager.contentsOfDirectory(at: cloudDocumentContainerUrl,
-                                                                   includingPropertiesForKeys: nil,
-                                                                   options: [])
-                    let fileURL = cloudDocumentContainerUrl.appendFile(name: fileName, fileExtension: fileExtension)
-                    if !urls.contains(fileURL) || force {
-                        let encoder = JSONEncoder()
-                        encoder.outputFormatting = .prettyPrinted
-                        let encodedContent = try encoder.encode(content)
-                        let created = fileManager.createFile(atPath: fileURL.path,
-                                                             contents: encodedContent, attributes: nil)
-                        completion(created, nil)
-                    } else {
-                        completion(false, CloudDocsError.fileAllreadyExists)
-                    }
-                } catch {
-                    completion(false, error)
-                }
+            } else {
+                completion(false, CloudDocsError.fileAllreadyExists)
+                return
             }
         }
+    }
+
+    func listAllFileURLs(completion: @escaping (_ urls: [URL]?, _ error: Error?) -> Void) {
+        guard let cloudDocumentContainerUrl = fileManager.cloudDocumentContainerUrl else {
+            completion(nil, CloudDocsError.cloudDocumentFolderNotFound)
+            return
+        }
+        _listAllFileURLs(from: cloudDocumentContainerUrl, completion: completion)
     }
 }
 
@@ -133,6 +156,8 @@ extension CloudDocs.CloudDocsError: LocalizedError {
             return NSLocalizedString("Given file path could not be found", comment: "")
         case .cloudDocumentFolderNotFound:
             return NSLocalizedString("Could not find cloud document folder", comment: "")
+        case .urlsNotFound:
+            return NSLocalizedString("Could not find urls", comment: "")
         }
     }
 }
@@ -141,12 +166,33 @@ internal extension CloudDocs {
     func createFolderIfNotExists(from url: URL, completion: @escaping (_ error: Error?) -> Void) {
         if fileManager.fileExists(atPath: url.path, isDirectory: nil) {
             completion(nil)
-        } else {
+            return
+        }
+        do {
+            try fileManager.createDirectory(at: url, withIntermediateDirectories: true, attributes: nil)
+            completion(nil)
+            return
+        } catch {
+            completion(error)
+            return
+        }
+    }
+
+    func _listAllFileURLs(from url: URL, completion: @escaping (_ urls: [URL]?, _ error: Error?) -> Void) {
+        createFolderIfNotExists(from: url) { (error: Error?) in
+            if let error = error {
+                completion(nil, error)
+                return
+            }
             do {
-                try fileManager.createDirectory(at: url, withIntermediateDirectories: true, attributes: nil)
-                completion(nil)
+                let urls = try fileManager.contentsOfDirectory(at: url,
+                                                               includingPropertiesForKeys: nil,
+                                                               options: [])
+                completion(urls, nil)
+                return
             } catch {
-                completion(error)
+                completion(nil, error)
+                return
             }
         }
     }
